# waterwind-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WaterWind API",
        "description": "WaterWind – Smart Water Recommendation Platform REST API. Authenticate via Sanctum Bearer tokens.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/statistics": {
            "get": {
                "tags": [
                    "Statistics"
                ],
                "summary": "Get overall statistics",
                "operationId": "bb183471907939a23e238bdabe05aa9f",
                "responses": {
                    "200": {
                        "description": "Statistics retrieved"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/statistics/crops": {
            "get": {
                "tags": [
                    "Statistics"
                ],
                "summary": "Get crop usage statistics",
                "operationId": "8e23637db75c73564fa3cf226eba6f13",
                "responses": {
                    "200": {
                        "description": "Crop statistics retrieved"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/statistics/soils": {
            "get": {
                "tags": [
                    "Statistics"
                ],
                "summary": "Get soil usage statistics",
                "operationId": "30c73f9b6273784ae6d38be129556aeb",
                "responses": {
                    "200": {
                        "description": "Soil statistics retrieved"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/statistics/minerals": {
            "get": {
                "tags": [
                    "Statistics"
                ],
                "summary": "Get mineral usage statistics",
                "operationId": "4bfb628e56e832a9d99999cffff30541",
                "responses": {
                    "200": {
                        "description": "Mineral statistics retrieved"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/statistics/devices": {
            "get": {
                "tags": [
                    "Statistics"
                ],
                "summary": "Get device usage statistics",
                "operationId": "cd41b912a3f967574e7b5c1ab9e0b620",
                "responses": {
                    "200": {
                        "description": "Device statistics retrieved"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/statistics/monthly": {
            "get": {
                "tags": [
                    "Statistics"
                ],
                "summary": "Get monthly analysis statistics",
                "operationId": "dc92383131db3d68d8363edc68ac9da3",
                "responses": {
                    "200": {
                        "description": "Monthly statistics retrieved"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/statistics/dashboard": {
            "get": {
                "tags": [
                    "Statistics"
                ],
                "summary": "Get dashboard-ready statistics",
                "operationId": "1dbb6100c6be7e2460ba00469f3564ab",
                "responses": {
                    "200": {
                        "description": "Dashboard statistics retrieved"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Sanctum token as: Bearer {token}",
                "scheme": "bearer"
            }
        }
    },
...
}